Only remove image file if it was temporary

+++ kstars/indi/indiccd.cpp
@@ -1224,7 +1224,10 @@ void CCD::processBLOB(IBLOB* bp)
             imageViewer = new ImageViewer(getDeviceName(), KStars::Instance());
 
         imageViewer->loadImage(filename);
-        QFile::remove(filename);
+
+        // Remove file from disk if it is preview or not a regular frame (focus, guide..etc)
+        if (targetChip->isBatchMode() == false || targetChip->getCaptureMode() != FITS_NORMAL)
+            QFile::remove(filename);
     }
     // Unless we have cfitsio, we're done.
 #ifdef HAVE_CFITSIO
